💄 style: support model switch with chat header
